make fem/test compile again and remove some warnings
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4694 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/function/blockvectorfunction/blockvectorfunction.hh
@@ -423,7 +423,7 @@ public:
   //! return dof
   DofType& operator *()
   {
-    assert((count_ >=0) && (count_ < dofArray_->size()));
+    assert((count_ >=0) && (count_ < (int)dofArray_->size()));
     //return DofWrapperType::convert((*dofArray_)[count_],idx_);
     return ((*dofArray_)[count_][idx_]);
   }
@@ -431,7 +431,7 @@ public:
   //! return dof read only 
   const DofType& operator * () const
   {
-    assert((count_ >=0) && (count_ < dofArray_->size()));
+    assert((count_ >=0) && (count_ < (int)dofArray_->size()));
     //return DofWrapperType::convert((*dofArray_)[count_],idx_);
     return ((*dofArray_)[count_][idx_]);
   }
